Re-work query loading
Start re-work of use of the client. Keep closer the JSON. All based on
the JSON response, hiding the clients wrapping and name mangling, which
is just confusing, especially when testing in the GH explorer.

Pull queries out to files for loading. Easier to edit in seperate file
(tooling) and much easier to move back and forward with the explorer.

Prototype Query class to handle running queries, getting, paging data
as JSON and insulation from client details.

Make --json an option to any action, instead of a seperate (hacky debug)
action.

Scan class handles archived and all as options on the class.